Revert "refactor: url configs"
This reverts commit 11a37533659b983d6da394c226b4a06ca29d1c77.
