+++ backend/apps/integrationserver/tests.py
@@ -1,18 +1,8 @@
-from django.test import TestCase
-from django.test.utils import override_settings
-
+from utils.testing.base import ExtendedGraphQLTestCase
 from utils.testing.factories.users import IndokUserFactory
 
 
-@override_settings(ROOT_URLCONF="config.urls.production")
-class ProductionServerTestCase(TestCase):
-    def test_cypress_disallowed(self):
-        response = self.client.get("/test-session/")
-        self.assertEqual(response.status_code, 404)
-
-
-@override_settings(ROOT_URLCONF="config.urls.test")
-class IntegrationServerTestCase(TestCase):
+class IntegrationServerTestCase(ExtendedGraphQLTestCase):
     def setUp(self) -> None:
         super().setUp()
         self.indok_user = IndokUserFactory(
@@ -29,5 +19,11 @@ class IntegrationServerTestCase(TestCase):
         )
 
     def test_cypress_allowed(self):
-        response = self.client.get("/test-session/")
-        self.assertEqual(response.status_code, 200)
+        with self.settings(URL_ROOTCONF="config.urls.test"):
+            response = self.client.get("/test-session/")
+            self.assertEqual(response.status_code, 200)
+
+    def test_cypress_disallowed(self):
+        with self.settings(URL_ROOTCONF="config.urls.production"):
+            response = self.client.get("/test-session/")
+            self.assertEqual(response.status_code, 404)
