+++ backend/config/urls/test.py
@@ -1,5 +1,5 @@
-from config.views import test
-from graphene_django.views import GraphQLView
+from config import views
+from config.views import CustomGraphQLView
 from django.conf import settings
 from django.urls import path
 from django.views.decorators.csrf import csrf_exempt
@@ -7,6 +7,6 @@ from django.views.decorators.csrf import csrf_exempt
 from .base import urlpatterns
 
 urlpatterns += [
-    path(settings.GRAPHQL_URL, GraphQLView.as_view(graphiql=True)),
-    path("test-session/", csrf_exempt(test.test_session)),
+    path(settings.GRAPHQL_URL, CustomGraphQLView.as_view(graphiql=True)),
+    path("/test-session/", csrf_exempt(views.test_session)),
 ]
