+++ backend/config/views.py
@@ -1,3 +1,6 @@
+from django.contrib.auth import get_user_model, login
+from django.http import HttpRequest, JsonResponse
+from django.middleware.csrf import get_token
 from graphene_django.views import GraphQLView
 from sentry_sdk.api import start_transaction
 
@@ -21,3 +24,20 @@ class CustomGraphQLView(GraphQLView):
                 operation_name,
                 show_graphiql=show_graphiql,
             )
+
+
+def test_session(request: HttpRequest):
+    print(request)
+    try:
+        login(
+            request,
+            get_user_model().objects.get(username="eva_student"),
+            backend="django.contrib.auth.backends.ModelBackend",
+        )
+        return JsonResponse({"username": request.user.username})
+    except get_user_model().DoesNotExist:
+        return JsonResponse({"username": None})
+
+
+def csrf(request: HttpRequest):
+    return JsonResponse({"csrfToken": get_token(request)})
