+++ backend/requirements/production.txt
@@ -1,3 +1,1 @@
 -r base.txt
-
-sentry-sdk==1.5.8
